# docs/partners/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/partner/articles/{id}": {
      "get": {
        "operationId": "PartnerController_article",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/PartnerContentDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "partner": []
          }
        ],
        "tags": [
          "Partner"
        ]
      }
    },
...
    "/v1/partner/articles/{id}/comments": {
      "get": {
        "operationId": "PartnerController_articleComments",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "pagination"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PartnerArticleCommentDto"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/PartnerPaginationDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "partner": []
          }
        ],
        "tags": [
          "Partner"
        ]
      }
    },
...
  },
  "info": {
    "title": "Men of Hunger Partner API",
    "description": "",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "partner": {
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "type": "http"
      }
    },
    "schemas": {
...
      "PartnerPaginationDto": {
        "type": "object",
        "properties": {
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "nextCursor"
        ]
      },
      "PartnerProfileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "bio": {
            "type": "string",
            "nullable": true
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "canonicalUrl": {
            "type": "string"
          },
          "accountKind": {
            "type": "string",
            "enum": [
              "person",
              "page"
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "counts": {
            "type": "object",
            "properties": {
              "followers": {
                "type": "integer"
              },
              "following": {
                "type": "integer"
              }
            }
          }
        },
        "required": [
          "id",
          "username",
          "name",
          "bio",
          "avatarUrl",
          "canonicalUrl",
          "accountKind",
          "createdAt"
        ]
      },
...
      "PartnerSourceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "url"
        ]
      },
      "PartnerEngagementDto": {
        "type": "object",
        "properties": {
          "boosts": {
            "type": "number",
            "description": "Visible boosts from permitted, non-banned accounts."
          },
          "comments": {
            "type": "number",
            "description": "Visible direct replies for posts, visible comments for articles."
          },
          "uniqueViewers": {
            "type": "number",
            "description": "Recorded distinct signed-in viewers, filtered for permitted accounts; not reach."
          },
          "reactions": {
            "type": "number",
            "description": "Visible article reaction records; a member can use multiple reactions."
          }
        },
        "required": [
          "boosts",
          "comments",
          "uniqueViewers"
        ]
      },
      "PartnerMediaDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string"
          },
          "url": {
            "type": "string",
            "nullable": true
          },
          "alt": {
            "type": "string",
            "nullable": true
          },
          "width": {
            "type": "number",
            "nullable": true
          },
          "height": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "kind",
          "url",
          "alt",
          "width",
          "height"
        ]
      },
      "PartnerContentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "post",
              "article"
            ]
          },
          "canonicalUrl": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "published"
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "editedAt": {
            "type": "string",
            "nullable": true
          },
          "body": {
            "type": "string"
          },
          "bodyFormat": {
            "type": "string",
            "enum": [
              "text",
              "html"
            ]
          },
          "publishedAt": {
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/PartnerSourceDto"
          },
          "engagement": {
            "$ref": "#/components/schemas/PartnerEngagementDto"
          },
          "title": {
            "type": "string"
          },
          "excerpt": {
            "type": "object",
            "nullable": true
          },
          "author": {
            "$ref": "#/components/schemas/PartnerProfileDto"
          },
          "media": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PartnerMediaDto"
            }
          }
        },
        "required": [
          "id",
          "kind",
          "canonicalUrl",
          "status",
          "createdAt",
          "editedAt",
          "body",
          "bodyFormat",
          "publishedAt",
          "source",
          "author",
          "media"
        ]
      },
      "PartnerArticleCommentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "articleId": {
            "type": "string"
          },
          "parentId": {
            "type": "string",
            "nullable": true
          },
          "body": {
            "type": "string"
          },
          "bodyFormat": {
            "type": "string",
            "enum": [
              "text"
            ]
          },
          "canonicalUrl": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "published"
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "editedAt": {
            "type": "string",
            "nullable": true
          },
          "author": {
            "$ref": "#/components/schemas/PartnerProfileDto"
          }
        },
        "required": [
          "id",
          "articleId",
          "parentId",
          "body",
          "bodyFormat",
          "canonicalUrl",
          "status",
          "createdAt",
          "editedAt",
          "author"
        ]
      }
    }
  }
}